requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==2.0.2
joblib==1.4.2
scikit-learn==1.5.2
python-dateutil==2.9.0.post0
pydantic==2.8.2
